separated api calls from main app
