+++ src/api/PostAPi.js
@@ -1,0 +1,47 @@
+import { useState, useEffect } from "react";
+import axios from "axios";
+
+function usePostApi() {
+  const [posts, setPosts] = useState([]);
+
+  useEffect(() => {
+    getPosts();
+  }, []);
+
+  const getPosts = () => {
+    axios
+      .get(`${process.env.REACT_APP_BACKEND_URL}/posts`)
+      .then((response) => {
+        console.log("response: ", response.data);
+        setPosts(response.data);
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  };
+
+  const addPost = (title, author, content) => {
+    axios
+      .post(`${process.env.REACT_APP_BACKEND_URL}/posts`, {
+        title,
+        author,
+        content,
+      })
+      .then((result) => {
+        console.log("result: ", result);
+        console.log(result.data);
+        const newPost = {
+          post_id: result.data.post.post_id,
+          title: result.data.post.title,
+          author: result.data.post.author,
+          content: result.data.post.content,
+        };
+        setPosts([...posts, newPost]);
+      })
+      .catch((error) => console.log(error.response.data));
+  };
+
+  return <>{addPost}</>;
+}
+
+export default usePostApi;
