+++ src/components/home.js
@@ -1,16 +1,20 @@
 import React from "react";
- 
-function Home() {
-    return (
-        <div>
-          <h1>Find the perfect caretaker for your fluffy friend here!</h1>
-         
-          <div className="content">
-             
-          </div>
-        </div>
-    );
-  }
+import NewPostForm from "./NewPostForm";
+import addPost from "../api/PostAPi";
 
- 
-  export default Home;
+function Home() {
+  return (
+    <div>
+      <h1>Find the perfect caretaker for your fluffy friend here!</h1>
+
+      <div className="content">
+        <section className="newPost">
+          <h1>Add Post:</h1>
+          <NewPostForm onAddPostCallback={addPost} />
+        </section>
+      </div>
+    </div>
+  );
+}
+
+export default Home;
